Renamed shared collection notice alert to container
