+++ prime-angular-frontend/src/app/modules/enrolment/shared/components/enrolment-collection-notice/enrolment-collection-notice.component.html
@@ -1,5 +1,5 @@
-<app-collection-notice-alert [isFull]="isFull"
-                             (accepted)="onAccept()">
+<app-collection-notice-container [isFull]="isFull"
+                                 (accepted)="onAccept()">
   <p>
     The personal information you provide to the PRIME application is collected by the British Columbia Ministry of
     Health under the authority of s. 26(a) and 26(c) of the Freedom of Information and Protection of Privacy Act
@@ -10,4 +10,4 @@
     Health Insurance BC toll free at <app-prime-phone></app-prime-phone> or <app-prime-support-email>
     </app-prime-support-email>.
   </p>
-</app-collection-notice-alert>
+</app-collection-notice-container>
