+++ prime-angular-frontend/src/app/modules/enrolment/shared/components/enrolment-collection-notice/enrolment-collection-notice.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { ICollectionNoticeAlert } from '@shared/components/collection-notice-alert/collection-notice-alert.component';
+import { ICollectionNoticeAlert } from '@shared/components/collection-notice-container/collection-notice-container.component';
 
 @Component({
   selector: 'app-enrolment-collection-notice',
