+++ prime-angular-frontend/src/app/modules/site-registration/pages/confirmation/confirmation.component.html
@@ -1,6 +1,6 @@
 <app-page>
 
-  <app-collection-notice-alert></app-collection-notice-alert>
+  <app-collection-notice-container></app-collection-notice-container>
 
   <app-page-header>PharmaNet Enrolment</app-page-header>
