+++ prime-angular-frontend/src/app/modules/site-registration/shared/components/site-collection-notice/site-collection-notice.component.html
@@ -1,5 +1,5 @@
-<app-collection-notice-alert [isFull]="isFull"
-                             (accepted)="onAccept()">
+<app-collection-notice-container [isFull]="isFull"
+                                 (accepted)="onAccept()">
   <p>
     The information collected in this form will be used by the Ministry of Health to permit and monitor PharmaNet
     access at this site.
@@ -13,4 +13,4 @@
   <p>
     Health Insurance BC toll free at <app-prime-phone></app-prime-phone>.
   </p>
-</app-collection-notice-alert>
+</app-collection-notice-container>
