+++ prime-angular-frontend/src/app/shared/components/collection-notice-container/collection-notice-container.component.html
